Initial integer type support. qfcc /is/ partially broken when it comes to integer constants and float function args/return values.
pr_comp.h:
	o  add the integer opcodes to pr_opcode_e
pr_edict.c:
	o  add "quaternion" and "integer" to type_name[]
	o  support quatnernion and integers types when printing values
	o  support the integer opcodes when bounds checking
pr_exec.c
	o  enable the integer opcodes
pr_opcode:
	o  add the integer opcodes to the opcode table
	o  logical operators all result in an integer rather than a value
expr.h:
	o  rename int_val to integer_val
qfcc.h:
	o  kill another magic number
expr.c:
	o  move the opcode to string conversion out of type_mismatch and into
	   get_op_string
	o  rename int_val to integer_val
	o  general integer type support.
	o  generate an internal comipiler error for null opcodes rather than
	   segging.
pr_imm.c:
	o  rename int_val to integer_val
	o  support integer constants, converting to float when needed.
pr_lex.c:
	o  magic number death and support quaternions and integers in type_size[]
qc-lex.l
	o  rename int_val to integer_val
	o  support quaternion and integer type keywords
qc-parse.y:
	o  rename int_val to integer_val
	o  use binary_expr instead of new_binary_expr for local initialized
	   variables
builtins.c:
	o  rename int_val to integer_val
	o  fix most (all?) of the INT related FIXMEs
defs.qc:
	o  use integer instead of float where it makes sense
main.c:
	o  read_result is now integer rather than float
main.qc:
	o  float -> integer where appropriate
	o  new test for int const to float arg

+++ tools/qwaq/main.c
@@ -22,7 +22,7 @@ void BI_Init (progs_t *progs);
 extern char *type_name[];
 extern cvar_t *developer;
 
-extern float *read_result; //FIXME: eww
+extern int *read_result; //FIXME: eww
 
 int
 main ()
@@ -101,7 +101,7 @@ main ()
 	}
 	printf ("\n");
 #endif
-	read_result = (float*)PR_GetGlobalPointer (&progs, "read_result");
+	read_result = (int*)PR_GetGlobalPointer (&progs, "read_result");
 	main_func = PR_GetFunctionIndex (&progs, "main");
 	PR_ExecuteProgram (&progs, main_func);
 	return G_FLOAT (&progs, OFS_RETURN);
